refactor(vault): extract canvas and bases file-type modules
Move listCanvasFiles/readCanvasFile/writeCanvasFile/updateCanvasFile plus
the canvas size/count caps and private helpers into src/lib/canvas.ts, and
listBaseFiles/readBaseFile into the existing src/lib/bases.ts (which already
owned MAX_BASE_FILE_BYTES). Both depend only on vault-fs; vault.ts re-exports
every moved symbol so the ./vault.js API is unchanged.
